Extend support to more views' XML attributes (v1)
The XML attributes whose related methods expect as values, (view) class constants in either the form "ATTRIBUTE_VALUE" or in the form "ATTRIBUTE_NAME" + "_" + "ATTRIBUTE_VALUE" are now supported. This was done by adding the function "toUnderscore" and minor tweaks in "getInteger".

The new supported attributes are:

[View class]
-  accessibilityLiveRegion
- drawingCacheQuality
- importantForAccessibility
- layoutDirection
- textAlignment
- textDirection

[TextView class]
- breakStrategy
- hyphenationFrequency

[GridLayout]
- alignmentMode

There might be more attributes supported but the previous are the ones I could test.