Restructure ViewInflater code
Most of the code related with the handling of the views' attributes was placed inside of the ViewAttributesHelper inner class.

This helper class is (mostly) used indirectly by using reflection to call its functions. In order to know which function should be called given some attribute, the mXmlAttrs HashMap is used to get some information about the attribute.

mXmlAttrs is a map where each key corresponds to a view attribute, and each value is a EnumMap that gives some information about an attribute.

The information given by the value (in this order) is:
- The name of the method of ViewAttributesHelper in case the attribute needs direct handling.
- The correct name of the setting method offered by Android (normally it can be obtained using some heuristics like "set" + PascalCase(attrName) but in many cases this doesn't apply).
- A modifier for the value input by the user (this is good to handle dimension values, ids and colors without having to write a dedicated function for the respective attribute).
- A class to help setDynamicProperty to retrieve the correct constants matching the values accepted by a given attribute (most are defined inside the View class, but in several cases are not).

Not all the information is required for each enum map (in fact, by definition no attribute should need all the values in for its enum map). Each mXmlAttrs value (i.e., the enum map) is initialized "on the fly" by calling mapAttrInfo. This has been the cleanest way I've come up with to do this.