Extend support to more views' XML attributes (v3)
In order to support new view attributes, several functions were written/modified in ViewInflater inner class ViewAttributeHelper.

Additionally, now each attribute in the helper hash map can specify a string to prefix the value in order to help getInteger to retrieve the corresponding constant value (the place to specify this string in the enum map, is in the one corresponding to CONSTANTS). Therefore, now it's possible to specify either a class where to look for the constants or a prefix to help retrieving them.

The view attributes that get support in this patch are:

View:
- backgroundTint
- paddingBottom
- paddingEnd
- paddingLeft
- paddingRight
- paddingStart
- paddingTop
- requiresFadingEdge

ImageView:
- tint

TextView:
- autoText
- capitalize