Extend support to more views' XML attributes (v4)
The mConflictiveAttrs HashMap introduced in the previous patch was modified to hold another HashMap as a value, since it didn't relate the conflictive attributes to a specific view, but to all. Now it does.

In order to support more view attributes, several dedicated functions were written in the ViewAttributesHelper inner class.

The view attributes that get support in this patch are:

TextView:
- drawableBottom
- drawableEnd
- drawableLeft
- drawablePadding
- drawableRight
- drawableStart
- drawableTop
- ellipsize
- fontFamily
- imeActionId
- imeActionLabel