Fix bugs in UiFacade related to dialogs
The dialogCreateHorizontalProgress and dialogTimePicker methods in UiFacade.java  had a minor bug in which the application had to be closed perforce if dialogDismiss wasn't called explicitly after invoking them.

dialogGetSelectedItems didn't return the selected items but the ones that were selected by default with dialogSetSingleChoiceItems/dialogSetMultiChoiceItems. This happened because unlike dialogGetResponse (which uses a Future object to get results from dialogs), it called the Set object that contained the information about the selected items. The problem was, that this Set object was read immediately regardless the dialog had dismissed or not. Now it uses the same Future object as dialogGetResponse to get the results, but treated to return a Set<Integer> object as dialogGetSelectedItems was originally thought.