Fix issues in UiFacade related to webViewShow
In HtmlActivityTask, the methods that are to be accessed from JavaScript have now the corresponding annotation (@JavascriptInterface) as it is mandatory from API 17.

Since UiFacade.webViewShow didn't work with external URLs (it was only able to load local html files),  now it delegates this responsibility to a external browser (local html files are still inside the web view).

Another issue that webViewShow had, was that it was impossible to go back to SL4A once it was executed (the application had to be closed by force). For this,  I implemented MyWebView class extending the WebView class. In this subclass, the onKeyDown method is overridden so that when the hardware back button is pressed, the previous screen is called (allowing us to return to SL4A without having to close it by force).