Fix layout_* attributes and parseTheme bugs in ViewInflater
With the introduction of the xXmlAttrs helper map, a bug was also introduced: many of the attributes whose name started with "layout_" weren't recognized anymore.This is due to the nature of the map: each attribute that needs to be mapped, has to be there. The problem is that the set of attributes that start by "layout_" is huge, so introducing all of them became a hassle. Looking for the prefix in setProperty may not be the most elegant solution, but it's the simplest one for this case.

parseTheme wasn't able to retrieve themes whose name contained '.' in xml, since these names as defined in R.style contain '_'. That was easily solved by replacing '.' by '_'.

PS: parseTheme query construction was modified to use a StringBuilder since it's more efficient.