Extend support for view attrs by recognizing more class constants
The map mXmlAttrs now can hold prefixes and suffixes for class constants. Now, not only the class holding the constants matching the attributes values can be given, but it's also possible to specify prefixes and suffixes that need to be appended in order to be recognized by getInteger.

Note: mXmlAttrs map initializer mapAttrInfo was modified in order to accept AttributeInfo-String pairs. This was done since having too many null values passed as parameters was messy and less readable. Now, we must pass a parameter specifying the type of info the attribute will have (from the AttributeInfo enum), and immediately after the corresponding info String. This as many times as the attribute needs it.